Upgraded to 3.7
Upgraded the module to be compatible with Fork CMS 3.7 or higher. You
might also want to change the download on the website to this version.
