--- backend/modules/twitter/info.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module>
-	<name>twitter</name>
-	<version>1.1.0</version>
-	<requirements>
-		<minimum_version>3.5.0</minimum_version>
-	</requirements>
-	<description>
-		<![CDATA[
-			Display your tweets on your website.
-		]]>
-	</description>
-	<authors>
-		<author>
-			<name><![CDATA[Bert Pattyn]]></name>
-			<url><![CDATA[http://www.dextrose.be]]></url>
-		</author>
-		<author>
-			<name><![CDATA[Dieter Vanden Eynde]]></name>
-			<url><![CDATA[http://www.dieterve.be]]></url>
-		</author>
-	</authors>
-	<events>
-		<event application="backend" name="after_add"><![CDATA[Triggered when a twitter widget is added.]]></event>
-		<event application="backend" name="after_delete"><![CDATA[Triggered when a twitter widget is deleted.]]></event>
-		<event application="backend" name="after_edit"><![CDATA[Triggered when a twitter widget is edited.]]></event>
-		<event application="backend" name="after_oauth"><![CDATA[Triggered when oauth authentication is complete.]]></event>
-		<event application="backend" name="after_saved_settings"><![CDATA[Triggered when the settings are saved.]]></event>
-	</events>
-	<cronjobs>
-		<cronjob minute="24" hour="*" day-of-month="*" month="*" day-of-week="*" action="get_tweets"><![CDATA[Looks for new tweets on the subscribed accounts.]]></cronjob>
-	</cronjobs>
-</module>
