+++ src/Frontend/Modules/Twitter/Config.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Frontend\Modules\Twitter;
+
+/*
+ * This file is part of Fork CMS.
+ *
+ * For the full copyright and license information, please view the license
+ * file that was distributed with this source code.
+ */
+
+use Frontend\Core\Engine\Base\Config as FrontendBaseConfig;
+
+/**
+ * This is the configuration-object.
+ *
+ * @author Bert Pattyn <bert@netlash.com>
+ * @author Tim van Wolfswinkel <tim@webleads.nl>
+ */
+final class Config extends FrontendBaseConfig
+{
+    /**
+     * The default action
+     *
+     * @var    string
+     */
+    protected $defaultAction = 'index';
+
+    /**
+     * The disabled actions
+     *
+     * @var    array
+     */
+    protected $disabledActions = array();
+}
